More Refactoring
