+++ app/app/src/main/java/com/example/myapplication/ui/myevents/MyEventsFragment.java
@@ -1,0 +1,35 @@
+package com.example.myapplication.ui.myevents;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.example.myapplication.databinding.FragmentMyeventsBinding;
+
+public class MyEventsFragment extends Fragment {
+
+    private FragmentMyeventsBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        MyEventsViewModel myEventsViewModel =
+                new ViewModelProvider(this).get(MyEventsViewModel.class);
+
+            binding = FragmentMyeventsBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+
+
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
