+++ app/app/src/main/java/com/example/myapplication/ui/myevents/MyEventsViewModel.java
@@ -1,14 +1,14 @@
-package com.example.myapplication.ui.gallery;
+package com.example.myapplication.ui.myevents;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class GalleryViewModel extends ViewModel {
+public class MyEventsViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public GalleryViewModel() {
+    public MyEventsViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is gallery fragment");
     }
