+++ app/app/src/main/java/com/example/myapplication/ui/registeredevents/RegisteredEventFragment.java
@@ -1,0 +1,34 @@
+package com.example.myapplication.ui.registeredevents;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.example.myapplication.databinding.FragmentRegisteredEventsBinding;
+
+public class RegisteredEventFragment extends Fragment {
+
+        private FragmentRegisteredEventsBinding binding;
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        RegisteredEventViewModel registeredEventViewModel =
+                new ViewModelProvider(this).get(RegisteredEventViewModel.class);
+
+            binding = FragmentRegisteredEventsBinding.inflate(inflater, container, false);
+        View root = binding.getRoot();
+
+        return root;
+    }
+
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        binding = null;
+    }
+}
