+++ app/app/src/main/java/com/example/myapplication/ui/registeredevents/RegisteredEventViewModel.java
@@ -1,14 +1,14 @@
-package com.example.myapplication.ui.slideshow;
+package com.example.myapplication.ui.registeredevents;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class SlideshowViewModel extends ViewModel {
+public class RegisteredEventViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public SlideshowViewModel() {
+    public RegisteredEventViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is slideshow fragment");
     }
