+++ app/app/src/main/res/layout/fragment_registered_events.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/back_gray"
-    tools:context=".ui.slideshow.SlideshowFragment">
+    tools:context=".ui.registeredevents.RegisteredEventFragment">
 
     <TextView
         android:id="@+id/text_slideshow"
