+++ app/app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.example.myapplication.ui.gallery.GalleryFragment"
+        android:name="com.example.myapplication.ui.myevents.MyEventsFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_myevents" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.example.myapplication.ui.slideshow.SlideshowFragment"
+        android:name="com.example.myapplication.ui.registeredevents.RegisteredEventFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_registered_events" />
 </navigation>
